// extrabeam-backend/src/common/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.4';
  };
  public: {
    Tables: {
// ...
      entreprise: {
        Row: {
          adresse_ligne1: string;
          adresse_ligne2: string | null;
          bic: string;
          code_postal: string | null;
          conditions_paiement: string;
          created_at: string;
          devise: string;
          email: string;
          iban: string;
          id: number;
          mention_tva: string;
          nom: string;
          pays: string | null;
          penalites_retard: string;
          prenom: string;
          siret: string;
          slug: string | null;
          statut_juridique: string;
          stripe_account_id: string | null;
          taux_horaire: number;
          telephone: string | null;
          tva_intracom: string | null;
          updated_at: string;
          user_id: string | null;
          ville: string | null;
        };
        Insert: {
          adresse_ligne1: string;
          adresse_ligne2?: string | null;
          bic: string;
          code_postal?: string | null;
          conditions_paiement?: string;
          created_at?: string;
          devise?: string;
          email: string;
          iban: string;
          id?: number;
          mention_tva?: string;
          nom: string;
          pays?: string | null;
          penalites_retard?: string;
          prenom: string;
          siret: string;
          slug?: string | null;
          statut_juridique?: string;
          stripe_account_id?: string | null;
          taux_horaire?: number;
          telephone?: string | null;
          tva_intracom?: string | null;
          updated_at?: string;
          user_id?: string | null;
          ville?: string | null;
        };
        Update: {
          adresse_ligne1?: string;
          adresse_ligne2?: string | null;
          bic?: string;
          code_postal?: string | null;
          conditions_paiement?: string;
          created_at?: string;
          devise?: string;
          email?: string;
          iban?: string;
          id?: number;
          mention_tva?: string;
          nom?: string;
          pays?: string | null;
          penalites_retard?: string;
          prenom?: string;
          siret?: string;
          slug?: string | null;
          statut_juridique?: string;
          stripe_account_id?: string | null;
          taux_horaire?: number;
          telephone?: string | null;
          tva_intracom?: string | null;
          updated_at?: string;
          user_id?: string | null;
          ville?: string | null;
        };
        Relationships: [];
      };
// ...
      missions: {
        Row: {
          client_id: string | null;
          contact_email: string;
          contact_name: string | null;
          contact_phone: string;
          created_at: string;
          devis_url: string | null;
          entreprise_id: number | null;
          etablissement: string;
          etablissement_adresse_ligne1: string | null;
          etablissement_adresse_ligne2: string | null;
          etablissement_code_postal: string | null;
          etablissement_pays: string | null;
          etablissement_ville: string | null;
          freelance_id: string | null;
          id: number;
          instructions: string | null;
          mode: Database['public']['Enums']['mission_mode'];
          status: Database['public']['Enums']['mission_status'];
        };
        Insert: {
          client_id?: string | null;
          contact_email: string;
          contact_name?: string | null;
          contact_phone: string;
          created_at?: string;
          devis_url?: string | null;
          entreprise_id?: number | null;
          etablissement: string;
          etablissement_adresse_ligne1?: string | null;
          etablissement_adresse_ligne2?: string | null;
          etablissement_code_postal?: string | null;
          etablissement_pays?: string | null;
          etablissement_ville?: string | null;
          freelance_id?: string | null;
          id?: number;
          instructions?: string | null;
          mode?: Database['public']['Enums']['mission_mode'];
          status?: Database['public']['Enums']['mission_status'];
        };
        Update: {
          client_id?: string | null;
          contact_email?: string;
          contact_name?: string | null;
          contact_phone?: string;
          created_at?: string;
          devis_url?: string | null;
          entreprise_id?: number | null;
          etablissement?: string;
          etablissement_adresse_ligne1?: string | null;
          etablissement_adresse_ligne2?: string | null;
          etablissement_code_postal?: string | null;
          etablissement_pays?: string | null;
          etablissement_ville?: string | null;
          freelance_id?: string | null;
          id?: number;
          instructions?: string | null;
          mode?: Database['public']['Enums']['mission_mode'];
          status?: Database['public']['Enums']['mission_status'];
        };
        Relationships: [
          {
            foreignKeyName: 'missions_client_id_fkey';
            columns: ['client_id'];
            isOneToOne: false;
            referencedRelation: 'profiles';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'missions_entreprise_id_fkey';
            columns: ['entreprise_id'];
            isOneToOne: false;
            referencedRelation: 'entreprise';
            referencedColumns: ['id'];
          },
        ];
      };
      profiles: {
        Row: {
          created_at: string | null;
          email: string | null;
          first_name: string | null;
          id: string;
          last_name: string | null;
          phone: string | null;
          role: string | null;
        };
        Insert: {
          created_at?: string | null;
          email?: string | null;
          first_name?: string | null;
          id: string;
          last_name?: string | null;
          phone?: string | null;
          role?: string | null;
        };
        Update: {
          created_at?: string | null;
          email?: string | null;
          first_name?: string | null;
          id?: string;
          last_name?: string | null;
          phone?: string | null;
          role?: string | null;
        };
        Relationships: [];
      };
      slots: {
        Row: {
          created_at: string;
          end: string | null;
          entreprise_id: number | null;
          id: number;
          mission_id: number | null;
          start: string | null;
          title: string | null;
        };
        Insert: {
          created_at?: string;
          end?: string | null;
          entreprise_id?: number | null;
          id?: number;
          mission_id?: number | null;
          start?: string | null;
          title?: string | null;
        };
        Update: {
          created_at?: string;
          end?: string | null;
          entreprise_id?: number | null;
          id?: number;
          mission_id?: number | null;
          start?: string | null;
          title?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: 'slots_entreprise_id_fkey';
            columns: ['entreprise_id'];
            isOneToOne: false;
            referencedRelation: 'entreprise';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'slots_mission_id_fkey';
            columns: ['mission_id'];
            isOneToOne: false;
            referencedRelation: 'missions';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      [_ in never]: never;
    };
    Enums: {
      facture_status: 'pending_payment' | 'paid' | 'canceled';
      mission_mode: 'freelance' | 'salarié';
      mission_status:
        | 'proposed'
        | 'validated'
        | 'pending_payment'
        | 'paid'
        | 'completed'
        | 'refused'
        | 'realized';
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};

type DatabaseWithoutInternals = Omit<Database, '__InternalSupabase'>;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  'public'
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema['Enums']
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;